Add site appearance and version switching system
- Add 'current' (Rhythmia) as default version alongside v1.0.0 (Discord UI) and v1.0.1 (Patreon UI)
- Add accent color customization (azure, purple, emerald, rose, amber) with CSS custom property injection
- Rewrite FloatingVersionSwitcher as combined settings panel with version + appearance sections
- Extract RhythmiaLobby from page.tsx into standalone component
- Make page.tsx a version-aware router that conditionally renders UI based on selected version
- Auto-restore version and accent preferences from localStorage on mount
